add choice

+++ bin/exec.sh
@@ -7,9 +7,14 @@ source ../conf/app.env
 export APP_VERSION=1.0
 APP_IMAGE=xianzixiang/`echo -n $PROJECT_NAME | awk '{print tolower($0)}'`:$APP_VERSION
 
-echo remove image ${APP_IMAGE} ...
+op=$1
+op=${op:=n}
 
-docker rmi $APP_IMAGE -f
+if [ $op != n ]; then
+
+    echo remove image ${APP_IMAGE} ...
+    docker rmi $APP_IMAGE -f
+fi
 
 echo running image [${APP_IMAGE}].....
